requirements.txt
et_xmlfile==2.0.0
Flask==3.1.1
gunicorn==25.2.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pdfplumber==0.11.6
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.1
xlrd==2.0.2
